+++ assets/unity/HippoRoyale_Build/manifest.webmanifest
@@ -1,0 +1,15 @@
+{
+    "name": "Four_Way_Soccer_Hippos",
+    "short_name": "Four_Way_Soccer_Hippos",
+    "start_url": "index.html",
+    "display": "fullscreen",
+    "background_color": "#231F20",
+    "theme_color": "#000",
+    "description": "",
+    "icons": [{
+      "src": "TemplateData/icons/unity-logo-dark.png",
+      "sizes": "144x144",
+      "type": "image/png",
+      "purpose": "any maskable"
+    }]
+  }
