+++ index.md
@@ -40,7 +40,7 @@ img.autoscale{
 <div class ="centerAlignObject">
     <h1>Hippo Soccer Royale</h1>
     <img src="HippoSoccerRoyaleAlpha_Pic.PNG"><br>
-<a href="/assets/unity/HippoRoyale_Build_0.0.0.1/index.html" class="button">Play</a>
+<a href="/assets/unity/HippoRoyale_Build/index.html" class="button">Play</a>
 </div>
 
 <div style="background-color:#444444; padding:10px;">
